Relocate to/from Hastus conversion objects.
